[VTPTools] Add vtp-move.py script
Moves anchors to new positions specified in a CSV file. The CSV format
is:

    anchor,x,y

With “anchor” being the anchor name (without the “MARK_” prefix), and
“x” and “y” being the new absolute position. The script will move all
mark anchors to the new position and adjust base anchors by the
difference between the old and new positions (this assume all mark
anchors are at the same position, otherwise the script will take the
difference of the first mark anchor).